Refactor message module into separate files
