+++ imessage-database/src/tables/messages/mod.rs
@@ -1,0 +1,9 @@
+/*!
+ Data structures and models used to parse and represent message data.
+*/
+
+pub use message::Message;
+
+pub(crate) mod body;
+pub mod message;
+pub mod models;
