+++ imessage-exporter/src/exporters/txt.rs
@@ -25,7 +25,7 @@ use imessage_database::{
     },
     tables::{
         attachment::Attachment,
-        messages::{BubbleType, Message},
+        messages::{models::BubbleType, Message},
         table::{Table, FITNESS_RECEIVER, ME, ORPHANED, YOU},
     },
     util::{
